requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework=3.12.2
sqlparse==0.3.0
whitenoise==5.0.1
django-cors-middleware==1.5.0
djangorestframework==3.11.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-related-admin==0.7.0
